refactor: Moved all dialogs in a sub module
