+++ src/gui/mod.rs
@@ -1,21 +1,7 @@
-pub mod load_dialog;
-pub mod process_dialog;
-pub mod prompt_save_dialog;
-pub mod save_dialog;
-pub mod type_selection_dialog;
-
-#[derive(Copy, Debug, Clone, Eq, PartialEq, Default)]
-pub enum DialogState<T>
-where
-    T: Clone,
-{
-    #[default]
-    Open, // Visible
-    Selected(T), // Selected
-    Cancelled,   // Cancelled
-}
-
-pub trait Dialog<T: Clone> {
-    fn state(&self) -> &DialogState<T>;
-    fn show(&mut self, ctx: &egui::Context);
-}
+pub mod dialogs;
+pub use dialogs::load_dialog;
+pub use dialogs::process_dialog;
+pub use dialogs::prompt_save_dialog;
+pub use dialogs::save_dialog;
+pub use dialogs::type_selection_dialog;
+pub use dialogs::{Dialog, DialogState};
